Move commodity CSV code to `input/commodity.rs`
